extend plot_plg to be able to add map backgrounds, add example to notebook
I did also refactor the exiting things a bit. This is not perfect yet, though.
Maybe a better solution would be to completely remove the map background stuff
from plot_lines and only have it in plot_plg because plot_lines could be a function
that is not exposed to the user anyway since plot_plg has the same functionality
plus the option to plot grid and points in addition...